+++ packages/cli/src/commands/view.ts
@@ -1,6 +1,7 @@
 import { Command } from 'commander'
 import loadConfig from '../config'
 import storage from '../storage'
+import * as api from '../api'
 
 const program = new Command('view')
 
@@ -8,17 +9,27 @@ export default () => {
   program
     .description('view configuration object')
     .argument('name', 'object to view ex. config, storage')
-    .action(async (name: string) => {
+    .option('-s, --symbol <symbol>', 'symbol to view cache data')
+    .option('-w, --window <window>', 'window size')
+    .action(async (name: string, opts) => {
       
       switch (name) {
         case 'config':
             const config = await loadConfig()
             console.dir(config, { depth: null, colors: true })
-          break;
+          break
 
         case 'storage':
             console.dir(storage.all, { depth: null, colors: true })
-          break;
+          break
+
+        case 'data':
+          const config2 = await loadConfig()
+          const data = api.cache(config2).getAll(opts.symbol, opts.window ?? 1440)
+
+          console.table(data)
+          break
+
         
         default:
           console.log('Invalid object to view. Please use config or storage.')